html/semantics/interactive-elements/the-dialog-element/remove-dialog-should-unblock-document.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body id="body">
    <dialog>
        This is a dialog
    </dialog>
    <input />
<script>
"use strict";
function testFocus(element, expectFocus) {
    element.focus();
    assert_equals(element === document.activeElement, expectFocus);
}

test(function() {
    var dialog = document.querySelector('dialog');
    dialog.showModal();

    var input = document.querySelector('input');
    testFocus(input, false);

    dialog.remove();
    testFocus(input, true);
}, "Test that removing dialog unblocks the document.");
</script>
</body>
</html>
