LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/remove-dialog-should-unblock-document-expected.txt

This is a dialog

FAIL Test that removing dialog unblocks the document. assert_equals: expected false but got true
